adding template bounds to kernels
