+++ cpp/src/examples/CMakeLists.txt
@@ -20,7 +20,6 @@ tx_add_exe(join_inmem_datagen_example)
 tx_add_exe(table_from_vectors_example)
 tx_add_exe(compute_example)
 tx_add_exe(groupby_example)
-tx_add_exe(test_group)
 
 
 #macro(tx_add_test_exe EXENAME)
